% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveSTClass.R
\docType{class}
\name{dfOrMtx-class}
\alias{dfOrMtx-class}
\alias{dfOrMtx}
\title{S4 Class for union of data.frame, matrix, and tibble}
\description{
An S4 class for a union of data.frame, matrix, and tibble
}
