% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_func.R
\docType{methods}
\name{generateCoorM}
\alias{generateCoorM}
\title{Generate patterns in simulations}
\usage{
generateCoorM(N_dup)
}
\arguments{
\item{N_dup}{how many duplicates to generate}
}
\value{
a list contaning nine D^2 matrix
}
\description{
Generate 9 matrix with patterns among spatial locations, this is factor genes in simulation.
}
