% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_func.R
\name{install_STexampleData}
\alias{install_STexampleData}
\title{Ensure STexampleData is installed}
\usage{
install_STexampleData()
}
\description{
This function checks if the STexampleData package is installed and installs it if necessary.
}
