% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveTransform.R
\docType{methods}
\name{InvWaveTrans}
\alias{InvWaveTrans}
\title{Inverse Two-Dimensional Discrete Wavelet Transform}
\usage{
InvWaveTrans(x, raws)
}
\arguments{
\item{x}{vectorized coefficients after wavelet transformation}

\item{raws}{original D^2 by P input}
}
\value{
a D^2 matrix
}
\description{
Performs Inverse wavelet transformation
}
