% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_func.R
\docType{methods}
\name{kOverA_ST}
\alias{kOverA_ST}
\title{k over A method for spatially resolved transcriptomics dataset}
\usage{
kOverA_ST(k = 5, A = 50)
}
\arguments{
\item{k}{The number of elements that have to exceed A.}

\item{A}{The value you want to exceed.}
}
\description{
Collect spatially resolved transcriptomics datasets in SpatialExperiment Bioconductor format using Visium_humanDLPFC_3_13,
then use kOverA select genes
}
