% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,waveST,missing-method}
\alias{plot,waveST,missing-method}
\title{plot function for waveST class}
\usage{
\S4method{plot}{waveST,missing}(x, k = 1, wave = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{waveST}}.}

\item{k}{which factor gene to visualize}

\item{wave}{whether use wavelet method}

\item{...}{Plot parameters forwarded.}
}
\value{
A plot object.
}
\description{
visualize the top factor genes and top 5 genes with high loadings on certain factor genes
}
