% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raws.R
\docType{data}
\name{raws}
\alias{raws}
\title{Cell by Gene ST data}
\format{
Y is a data frame with 4096 rows and 721 variables:
\describe{
\item{row}{each location}
\item{column}{each gene}
}
}
\usage{
raws
}
\description{
This is the D^2 by p matrix outputed by k=3 over A=7 method for spatially resolved transcriptomics dataset and input generation
}
\keyword{datasets}
