% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\docType{data}
\name{res}
\alias{res}
\title{Sample Spatial Transcriptomics Data}
\format{
A list with 2 elements:
\describe{
\item{viz}{A tibble with 4,992 rows and 2 columns, containing spatial coordinates}
\item{df}{A numeric matrix with 4,992 rows and 19 columns, containing gene expression values}
}
}
\source{
Generated from STexampleData package
}
\usage{
res
}
\description{
A dataset containing sample spatial transcriptomics data with coordinates and gene expression values.
}
\keyword{datasets}
