% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\docType{data}
\name{simulation_data}
\alias{simulation_data}
\title{Cell by Gene ST data}
\format{
Y is a data frame with 1024 rows and 300 variables:
\describe{
\item{row}{each location}
\item{column}{each gene}
}
}
\usage{
simulation_data
}
\description{
This D^2 by p matrix contains the synthetic data generated for simulation purposes in the method for spatially resolved transcriptomics datasets and input generation.
}
\keyword{datasets}
