% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveSTClass.R
\docType{class}
\name{waveST-class}
\alias{waveST-class}
\title{S4 Class for a waveST}
\description{
The waveST class stores the data and additional result after computing.
}
\details{
If \code{spatial} is not given, the class assumes the data is generated input: location(D^2)-by-gene(2).
The class set \code{input} = \code{data}, this is useful when we already have location(D^2)-by-gene(2)
matrix (like settings in simulations).\preformatted{      The content of \code{output} depends on whether use wavelet, and the decomposition methods.
      The  \code{output} is NA once created. The values will be updated once called \code{\link{decompse}}.
}
}
\note{
The decomposition in this package require a waveST input.
}
\section{Slots}{

\describe{
\item{data}{original data, cell(n)-by-gene(p) matrix, can be matrix/data.frame/tibble}
\item{spatial}{cell(n)-by-location(2) matrix, representing coordinates of each cell, see "details"}
\item{input}{location(D^2)-by-gene(2) matrix, representing gene expression over location}
\item{output}{a list containing SVD-like output: factors, diagonal matrix, loadings, fitted values, see "details"}
}
}

\examples{
wave = new("waveST", data = matrix(rnorm(6),2,3))
}
\author{
Zhuoyan Xu \email{zhuoyan.xu@wisc.edu}
}
