% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveTransform.R
\docType{methods}
\name{WaveTransCoefs}
\alias{WaveTransCoefs}
\title{wavelet transformation}
\usage{
WaveTransCoefs(x, wf = "d4", J = 4, thresholdMethod = NULL, tau = NULL)
}
\arguments{
\item{x}{length D^2 vector, vectorized matrix for DWT}

\item{wf}{name of the wavelet filter to use}

\item{J}{depth of the wavelet basis decomposition, must be a number less than or equal to log(min{M,N},2)}

\item{thresholdMethod}{wavelet shrinkage method used  "hybrid" thresholding or "manual" thresholding, see "details"}

\item{tau}{constant threshold when using manual thresholding}
}
\value{
a \code{tibble} with two columns, \code{level} and \code{coef}, see details
}
\description{
Performs a separable two-dimensional discrete wavelet transform (DWT) on a matrix of dyadic dimensions. Applying
indicated thresholds.
}
\details{
The coefficients after wavelet thresformation and thresholding is a list structure
containing the 3J+1 sub-matrices from the decomposition. The \code{coef} is the vectorized
coefficients.\preformatted{     The "hybrid" thresholding combined the SURE and universal thresholds. The manual thresholding set a constant threshold.
     see Section 3.2 in references.
}
}
\references{
G. P. Nason. Wavelet methods in statistics with R. Springer, 2008.
}
