% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveSTClass.R
\docType{methods}
\name{decompose}
\alias{decompose}
\title{waveST decomposition}
\usage{
decompose(
  waveST,
  wavemethod = c("raw", "wave"),
  decom_method = c("SVD", "EBMF"),
  K = NULL,
  bar = 500,
  wf = "d4",
  J = 5,
  thresholdMethod = "manual",
  tau = 0
)
}
\arguments{
\item{waveST}{an \code{waveST} class}

\item{wavemethod}{a vector showing whether whether we use wavelet transformation, "raw" means we decompose
directly without wavelet technique, "wave" means we apply wavelet transformation and shrinkage technique}

\item{decom_method}{Matrix decomposition method, "SVD" or "EBMF"}

\item{K}{The estimated number of factors, use "elbow" method to estimate it if not given}

\item{bar}{threshold for selecting number of factors, \code{K}, useless if \code{is already given}, see "details"}

\item{wf}{name of the wavelet filter to use}

\item{J}{depth of the wavelet basis decomposition, must be a number less than or equal to log(min{M,N},2)}

\item{thresholdMethod}{wavelet shrinkage method used  "hybrid" thresholding or "manual" thresholding, see "details"}

\item{tau}{constant threshold when using manual thresholding}
}
\value{
a \code{\link{waveST}} object with \code{output} updated, see "details"
}
\description{
Decompose input in \code{\link{waveST}} object with wavelet transformation and threshold
}
\details{
If the number of factors \code{K} is not given, the use \code{bar} to select \code{K}, it
conduct SVD first,only keep the singular values larger than \code{bar}, the number of
kept values are selected \code{K}.\preformatted{     The "hybrid" thresholding or "manual" thresholding, see \code{\link{WaveTransCoefs}}.
     
     The output contains:
     \\enumerate\{
     \\item \code{f} factor matrix, representing factor genes
     \\item \code{D} middle diagonal matrix 
     \\item \code{l} loading matrix
     \\item \code{recon} The fitted reconstruction matrix, estimated data
     \}
     
     If wavelet method is used, \code{f}, \code{D}, \code{l} cooresponding to the decomposition 
     result for coefficient matrix. \code{recon} is still the fitted matrix of original input.
}
}
\examples{
res = kOverA_ST(k = 3,A = 7)
viz = res$viz; df = res$df
wave = waveST(data = df[,1:5], spatial = viz)
wave = decompose(wave, "raw", "SVD", K = 5)
}
