% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveSTClass.R
\docType{methods}
\name{waveST}
\alias{waveST}
\title{waveST Constructor}
\usage{
waveST(data, spatial = NA, input = NA, level = 6)
}
\arguments{
\item{data}{an instance of \code{data.frame}, \code{matrix}, or \code{tibble}, representing riginal data, cell(n)-by-gene(p) matrix,}

\item{spatial}{cell(n)-by-location(2) matrix, representing coordinates of each cell, see "details"}

\item{input}{location(D^2)-by-gene(2) matrix, representing gene expression over location}

\item{level}{the scale for the gene expression image after input generation, the size of images for input generation should be 2^level}
}
\value{
a \code{\link{waveST}} object
}
\description{
Create a \code{\link{waveST}} object from an \code{data.frame}, \code{matrix}, or \code{tibble}.
}
\details{
If \code{spatial} is not given, the class assumes the data is generated input: location(D^2)-by-gene(2).
The class set \code{input} = \code{data}, this is useful when we already have location(D^2)-by-gene(2)
matrix (like settings in simulations).\preformatted{     The content of \code{output} depends on whether use wavelet, and the decomposition methods.
     The  \code{output} is NA once created. The values will be updated once called \code{\link{decompse}}.
}
}
\examples{
res = kOverA_ST(k = 3,A = 7)
viz = res$viz; df = res$df
wave = waveST(data = df, spatial = viz)
}
